requirements.txt
bcrypt==3.1.6
beautifulsoup4>=4.7.1
colorlog>=4.0.2
coverage>=4.5.2
Django==2.1.7
django-anymail>=5.0
django-compressor>=2.2
django-debug-toolbar>=1.11
django-downloadview>=1.9
django-environ==0.4.5
django-heroku==0.3.1
django-hijack>=2.1.10
django-hijack-admin>=2.1.10
django-htmlmin>=0.10.0
django-unused-media>=0.1.13
django-webtest>=1.9.4
factory-boy>=2.11.1
gunicorn==19.9.0
html5lib>=1.0.1
psycopg2-binary==2.7.7
PyYAML>=3.13
raven>=6.10.0
requests>=2.21.0
